requirements.txt
absl-py==0.15.0
# adbutils==0.11.0
aiofiles==0.7.0
altunityrunner==1.6.6
annotated-types==0.5.0
anyio==3.7.1
apkutils2==1.0.0
Appium-Python-Client==1.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
behave==1.2.6
bleach==5.0.1
cached-property==1.5.2
cachetools==5.2.0
certifi==2021.5.30
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.0.4
cigam==0.0.3
click==8.0.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.0.9
defusedxml==0.7.1
Deprecated==1.2.12
deprecation==2.1.0
Django==3.1.3
docutils==0.17.1
driver==0.1.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.0.0
facebook-wda==1.4.3
fastapi==0.103.1
fastjsonschema==2.16.1
fb-idb==1.1.2
filelock==3.0.12
flatbuffers==1.12
fonttools==4.37.1
fst-pso==1.8.1
future==0.18.2
FuzzyTM==2.0.5
gast==0.4.0
gensim==4.3.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
grpclib==0.4.1
guesslang==2.2.1
h2==4.0.0
h5py==3.1.0
hpack==4.0.0
huggingface-hub==0.11.0
hyperframe==6.0.1
idna==3.2
importlib-metadata==4.12.0
iniconfig==2.0.0
install==1.3.5
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jax==0.4.8
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
JPype1==1.4.0
jsonschema==4.15.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.12.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kitchen==1.2.6
kiwisolver==1.4.4
konlpy==0.6.0
libclang==16.0.0
login==0.0.6
loguru==0.5.3
logzero==1.7.0
lxml==4.6.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
miniful==0.0.6
mistune==2.0.4
ml-dtypes==0.1.0
multidict==5.1.0
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
nose==1.3.7
notebook==6.4.12
numpy==1.19.5
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.4.1
pandocfilters==1.5.0
parse==1.19.0
parse-type==0.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pluggy==1.0.0
progress==1.6
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pure-python-adb==0.3.0.dev0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyelftools==0.27
pyFUME==0.2.25
Pygments==2.13.0
pyparsing==3.0.0b3
PyQt3D==5.15.4
PyQt5==5.15.4
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
PyQtChart==5.15.4
PyQtDataVisualization==5.15.4
PyQtNetworkAuth==5.15.4
PyQtPurchasing==5.15.4
PyQtWebEngine==5.15.4
pyrsistent==0.18.1
pytagcloud==0.3.5
pytest==7.3.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
regex==2022.3.2
requests==2.28.1
requests-oauthlib==1.3.1
retry==0.9.2
robotframework==4.1
robotframework-appiumlibrary==1.6.2
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
selenium==3.141.0
self==2020.12.3
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
simpful==2.9.0
six==1.15.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.0
starlette==0.27.0
TBB==0.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.63.0
traitlets==5.3.0
transformers==4.24.0
treelib==1.6.1
tweepy==4.10.1
typing-extensions==3.7.4.3
uiautomator2==2.16.7
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
weditor==0.6.4
Werkzeug==2.2.2
whichcraft==0.6.1
widgetsnbextension==4.0.3
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.8.1
